0.19.3 new package for Armors + refactoring
