+++ wfrp/src/main/java/com/warhammer/wfrp/eyes/models/EyeColorInterface.java
@@ -1,4 +1,4 @@
-package com.warhammer.wfrp.model.eyes;
+package com.warhammer.wfrp.eyes.models;
 
 public interface EyeColorInterface {
