+++ wfrp/src/main/java/com/warhammer/wfrp/eyes/models/HalflingEyeColor.java
@@ -1,4 +1,4 @@
-package com.warhammer.wfrp.model.eyes;
+package com.warhammer.wfrp.eyes.models;
 
 import jakarta.persistence.*;
 import lombok.Getter;
@@ -7,8 +7,8 @@ import lombok.Setter;
 @Getter
 @Setter
 @Entity
-@Table(name = "DWARF_EYE_COLOR")
-public class DwarfEyeColor implements EyeColorInterface {
+@Table(name = "HALFLING_EYE_COLOR")
+public class HalflingEyeColor implements EyeColorInterface {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
