+++ wfrp/src/main/java/com/warhammer/wfrp/eyes/models/HighElfEyeColor.java
@@ -1,4 +1,4 @@
-package com.warhammer.wfrp.model.eyes;
+package com.warhammer.wfrp.eyes.models;
 
 import jakarta.persistence.*;
 import lombok.Getter;
@@ -7,8 +7,8 @@ import lombok.Setter;
 @Getter
 @Setter
 @Entity
-@Table(name = "HALFLING_EYE_COLOR")
-public class HalflingEyeColor implements EyeColorInterface {
+@Table(name = "HIGH_ELF_EYE_COLOR")
+public class HighElfEyeColor implements EyeColorInterface {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
