+++ wfrp/src/main/java/com/warhammer/wfrp/eyes/repositories/DwarfEyeColorRepository.java
@@ -1,6 +1,6 @@
-package com.warhammer.wfrp.repository.eyes;
+package com.warhammer.wfrp.eyes.repositories;
 
-import com.warhammer.wfrp.model.eyes.DwarfEyeColor;
+import com.warhammer.wfrp.eyes.models.DwarfEyeColor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DwarfEyeColorRepository extends JpaRepository<DwarfEyeColor, Integer> {
