+++ wfrp/src/main/java/com/warhammer/wfrp/eyes/repositories/HalflingEyeColorRepository.java
@@ -1,6 +1,6 @@
-package com.warhammer.wfrp.repository.eyes;
+package com.warhammer.wfrp.eyes.repositories;
 
-import com.warhammer.wfrp.model.eyes.HalflingEyeColor;
+import com.warhammer.wfrp.eyes.models.HalflingEyeColor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface HalflingEyeColorRepository extends JpaRepository<HalflingEyeColor, Integer> {
