+++ wfrp/src/main/java/com/warhammer/wfrp/eyes/repositories/HighElfEyeColorRepository.java
@@ -1,6 +1,6 @@
-package com.warhammer.wfrp.repository.eyes;
+package com.warhammer.wfrp.eyes.repositories;
 
-import com.warhammer.wfrp.model.eyes.HighElfEyeColor;
+import com.warhammer.wfrp.eyes.models.HighElfEyeColor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface HighElfEyeColorRepository extends JpaRepository<HighElfEyeColor, Integer> {
