+++ wfrp/src/main/java/com/warhammer/wfrp/eyes/repositories/HumanEyeColorRepository.java
@@ -1,6 +1,6 @@
-package com.warhammer.wfrp.repository.eyes;
+package com.warhammer.wfrp.eyes.repositories;
 
-import com.warhammer.wfrp.model.eyes.HumanEyeColor;
+import com.warhammer.wfrp.eyes.models.HumanEyeColor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface HumanEyeColorRepository extends JpaRepository<HumanEyeColor, Integer> {
