+++ wfrp/src/main/java/com/warhammer/wfrp/eyes/repositories/WoodElfEyeColorRepository.java
@@ -1,6 +1,6 @@
-package com.warhammer.wfrp.repository.eyes;
+package com.warhammer.wfrp.eyes.repositories;
 
-import com.warhammer.wfrp.model.eyes.WoodElfEyeColor;
+import com.warhammer.wfrp.eyes.models.WoodElfEyeColor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface WoodElfEyeColorRepository extends JpaRepository<WoodElfEyeColor, Integer> {
