+++ wfrp/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-server.port=8080
+server.port=8888
 
 ## MySQL
 spring.datasource.url=jdbc:mysql://localhost:3306/warhammer
@@ -8,7 +8,7 @@ spring.sql.init.mode=always
 spring.jpa.defer-datasource-initialization=true
 
 ## Hikari connection
-spring.datasource.hikari.connection-timeout=20000
+spring.datasource.hikari.connection-timeout=50000
 spring.datasource.hikari.maximum-pool-size=5
 
 ## Hibernate
